♻️ JanggiBoard(장기판 보여주기)와 Referee(장기판 업데이트) 역할 분리
